+++ src/Service/SessionManagerFactory.php
@@ -18,6 +18,7 @@ use Zend\Session\Container;
 use Zend\Session\SaveHandler\SaveHandlerInterface;
 use Zend\Session\SessionManager;
 use Zend\Session\Storage\StorageInterface;
+use Zend\Session\ManagerInterface;
 
 class SessionManagerFactory implements FactoryInterface
 {
@@ -124,7 +125,16 @@ class SessionManagerFactory implements FactoryInterface
             }
         }
 
-        $manager = new SessionManager($config, $storage, $saveHandler, $validators, $options);
+        $managerClass = class_exists($requestedName) ? $requestedName : SessionManager::class;
+        if (! is_subclass_of($managerClass, ManagerInterface::class)) {
+            throw new ServiceNotCreatedException(sprintf(
+                'SessionManager requires that the %s service implement %s',
+                $managerClass,
+                ManagerInterface::class
+            ));
+        }
+
+        $manager = new $managerClass($config, $storage, $saveHandler, $validators, $options);
 
         // If configuration enables the session manager as the default manager for container
         // instances, do so.