+++ test/Service/SessionManagerFactoryTest.php
@@ -9,9 +9,11 @@
 
 namespace ZendTest\Session\Service;
 
+use PHPUnit\Framework\TestCase;
 use Zend\EventManager\Test\EventListenerIntrospectionTrait;
 use Zend\ServiceManager\Config;
 use Zend\ServiceManager\ServiceManager;
+use Zend\ServiceManager\Exception\ServiceNotCreatedException;
 use Zend\Session\Config\ConfigInterface;
 use Zend\Session\Container;
 use Zend\Session\ManagerInterface;
@@ -21,12 +23,14 @@ use Zend\Session\SessionManager;
 use Zend\Session\Storage\ArrayStorage;
 use Zend\Session\Storage\StorageInterface;
 use Zend\Session\Validator;
+use ZendTest\Session\TestAsset\TestManager;
+use ZendTest\Session\TestAsset\TestSaveHandler;
 
 /**
  * @group      Zend_Session
  * @covers Zend\Session\Service\SessionManagerFactory
  */
-class SessionManagerFactoryTest extends \PHPUnit_Framework_TestCase
+class SessionManagerFactoryTest extends TestCase
 {
     use EventListenerIntrospectionTrait;
 
@@ -35,6 +39,8 @@ class SessionManagerFactoryTest extends \PHPUnit_Framework_TestCase
         $config = new Config([
             'factories' => [
                 ManagerInterface::class => SessionManagerFactory::class,
+                TestManager::class => SessionManagerFactory::class,
+                TestSaveHandler::class => SessionManagerFactory::class,
             ],
         ]);
         $this->services = new ServiceManager();
@@ -49,7 +55,7 @@ class SessionManagerFactoryTest extends \PHPUnit_Framework_TestCase
 
     public function testConfigObjectIsInjectedIfPresentInServices()
     {
-        $config = $this->getMock(ConfigInterface::class);
+        $config = $this->createMock(ConfigInterface::class);
         $this->services->setService(ConfigInterface::class, $config);
         $manager = $this->services->get(ManagerInterface::class);
         $test = $manager->getConfig();
@@ -68,7 +74,7 @@ class SessionManagerFactoryTest extends \PHPUnit_Framework_TestCase
 
     public function testFactoryWillInjectSaveHandlerIfPresentInServices()
     {
-        $saveHandler = $this->getMock(SaveHandlerInterface::class);
+        $saveHandler = $this->createMock(SaveHandlerInterface::class);
         $this->services->setService(SaveHandlerInterface::class, $saveHandler);
         $manager = $this->services->get(ManagerInterface::class);
         $test = $manager->getSaveHandler();
@@ -207,5 +213,17 @@ class SessionManagerFactoryTest extends \PHPUnit_Framework_TestCase
 
         $manager = $this->services->get(ManagerInterface::class);
         $this->assertAttributeSame([], 'validators', $manager);
+    }
+
+    public function testFactoryWillUseRequestedNameAsSessionManagerIfItImplementsManagerInterface()
+    {
+        $manager = $this->services->get(TestManager::class);
+        $this->assertInstanceOf(TestManager::class, $manager);
+    }
+
+    public function testFactoryWillRaiseServiceNotCreatedExceptionIfRequestedNameIsNotAManagerInterfaceSubclass()
+    {
+        $this->expectException(ServiceNotCreatedException::class);
+        $manager = $this->services->get(TestSaveHandler::class);
     }
 }